mindsyncaiagent/requirements.txt
# requirements.txt
langchain>=0.2.12
langgraph>=0.2.4
langchain-google-genai>=0.1.0
pydantic>=2.6
python-dotenv>=1.0.1
fastapi>=0.115
uvicorn>=0.30
python-dateutil>=2.8
pytest